+++ torch_frame/__init__.py
@@ -1,5 +1,5 @@
 from .stype import (stype, numerical, categorical, text_embedded,
-                    multicategorical)
+                    text_tokenized, multicategorical)
 from .data import TensorFrame
 from .typing import TaskType, DataFrame, NAStrategy
 from torch_frame.utils import save, load, cat  # noqa
@@ -16,6 +16,7 @@ __all__ = [
     'numerical',
     'categorical',
     'text_embedded',
+    'text_tokenized',
     'multicategorical',
     'TaskType',
     'NAStrategy',