+++ torch_frame/data/stats.py
@@ -42,6 +42,8 @@ class StatType(Enum):
             ]
         elif stype == torch_frame.text_embedded:
             return []
+        elif stype == torch_frame.text_tokenized:
+            return []
         elif stype == torch_frame.multicategorical:
             return [
                 StatType.MULTI_COUNT,